requirements.txt
Flask
Flask-Cors
protobuf==6.32.0
pycryptodome==3.20.0
requests
